+++ src/reducers/profile.js
@@ -43,7 +43,9 @@ export default (state = initialState, action) => {
     case "PROFILE_VOTE_CARD_ID_ACQUIRED":
       return { ...state, voteCardId: payload.voteCardId };
     case "PROFILE_SAVING_AVATAR":
-      return { ...state, isSavingAvatar: payload.isSaving };
+      return { ...state, isSavingAvatar: payload.isSaving, avatarError: null };
+    case "PROFILE_SAVE_AVATAR_ERROR":
+      return { ...state, avatarError: payload.error };
     case "PROFILE_VALIDATE_PROFILE": {
       return { ...state, validatingProfile: true };
     }
@@ -53,6 +55,7 @@ export default (state = initialState, action) => {
     case "SESSION_CLEAR_SESSION":
       return {
         ...state,
+        avatarError: null,
         currentUser: null,
         isFetching: false,
         isSaving: false,