+++ src/sagas/profile.js
@@ -1,4 +1,4 @@
-import { call, fork, put, spawn, select, takeLatest } from "redux-saga/effects";
+import { all, call, fork, put, spawn, select, takeLatest } from "redux-saga/effects";
 import { Alert } from "react-native";
 
 import { monitorUpload } from "./upload";
@@ -25,23 +25,27 @@ import {
 } from "../actions";
 
 import {
+  avatarSaveErrors,
   currentAuthToken,
   currentUser,
+  getChangePasswordErrors,
   getCurrentSigningPlip,
   isProfileComplete,
+  profileSaveErrors,
 } from "../selectors";
 
 import {
   User,
 } from "../models";
 
-import { isDev, isUnauthorized, log, logError } from "../utils";
+import { isDev, isUnauthorized, log, logError, isBlank } from "../utils";
 
 import Toast from "react-native-simple-toast";
 
 import locale from "../locales/pt-BR";
 import { blockBuilder } from "./crypto";
 import { validateLocalWallet } from "./wallet";
+import { changePasswordSaga } from "./password";
 
 function* saveMainProfile({ mobileApi, sessionStore, Crypto }) {
   yield takeLatest("PROFILE_SAVE_MAIN", function* ({ payload }) {
@@ -252,83 +256,90 @@ function* savePhoneProfile({ mobileApi, DeviceInfo }) {
   });
 }
 
+function* saveAvatarProfileSaga({ mobileApi, avatar = {}, oldAvatarURL, shouldNavigate }) {
+  try {
+    yield put(savingAvatar(true));
+
+    const authToken = yield select(currentAuthToken);
+    const avatarPayload = avatar && avatar.uri && avatar.uri !== oldAvatarURL ?
+      { ...avatar } :
+      { oldAvatarURL };
+
+    log(`avatar payload: ${JSON.stringify(avatarPayload)}`);
+
+    yield call(monitorUpload, mobileApi.saveAvatar(authToken, avatarPayload), function ({ type, payload: uploadPayload }) {
+      switch (type) {
+        case "UPLOAD_PROGRESS":
+          log(`upload progress ${uploadPayload}`);
+          break;
+        case "UPLOAD_FINISHED":
+          log("upload finished", uploadPayload)
+          break;
+        case "UPLOAD_FAILED":
+          throw uploadPayload;
+      }
+    });
+
+    const response = yield call(mobileApi.profile, authToken);
+    const user = User.fromJson(response.user);
+
+    yield put(updatedUserProfile({ user }));
+    yield put(savingAvatar(false));
+
+    if (shouldNavigate) {
+      yield put(profileStateMachine());
+    }
+
+    yield put(logEvent({ name: "completed_avatar" }));
+  } catch (e) {
+    logError(e, { tag: "saveAvatarProfile" });
+
+    yield put(savingAvatar(false));
+
+    if (shouldNavigate && isUnauthorized(e)) return yield put(unauthorized({ type: "reset" }));
+
+    yield put(saveAvatarError(e));
+  }
+}
+
 function* saveAvatarProfile({ mobileApi }) {
   yield takeLatest("PROFILE_SAVE_AVATAR", function* ({ payload }) {
     const { avatar = {}, oldAvatarURL, shouldNavigate } = payload;
 
-    try {
-      yield put(savingAvatar(true));
-
-      const authToken = yield select(currentAuthToken);
-      const avatarPayload = avatar && avatar.uri && avatar.uri !== oldAvatarURL ?
-        { ...avatar } :
-        { oldAvatarURL };
-
-      log(`avatar payload: ${JSON.stringify(avatarPayload)}`);
-
-      yield call(monitorUpload, mobileApi.saveAvatar(authToken, avatarPayload), function ({ type, payload: uploadPayload }) {
-        switch (type) {
-          case "UPLOAD_PROGRESS":
-            log(`upload progress ${uploadPayload}`);
-            break;
-          case "UPLOAD_FINISHED":
-            log("upload finished", uploadPayload)
-            break;
-          case "UPLOAD_FAILED":
-            throw uploadPayload;
-        }
-      });
-
-      const response = yield call(mobileApi.profile, authToken);
-      const user = User.fromJson(response.user);
-
-      yield put(updatedUserProfile({ user }));
-      yield put(savingAvatar(false));
-
-      if (shouldNavigate)
-        yield put(profileStateMachine());
-      else {
-        yield call([Toast, Toast.show], locale.avatarSaved);
-      }
-
-      yield put(logEvent({ name: "completed_avatar" }));
-    } catch (e) {
-      logError(e, { tag: "saveAvatarProfile" });
-
-      yield put(savingAvatar(false));
-
-      if (shouldNavigate && isUnauthorized(e)) return yield put(unauthorized({ type: "reset" }));
-
-      yield put(saveAvatarError(e));
-    }
+    yield call(saveAvatarProfileSaga, { mobileApi, avatar, oldAvatarURL, shouldNavigate });
   });
+}
+
+function* updateProfileSaga({ mobileApi, birthdate, name, zipCode, voteIdCard, shouldNavigate }) {
+  try {
+    yield put(savingProfile(true));
+
+    const authToken = yield select(currentAuthToken);
+    const response = yield call(mobileApi.updateProfile, authToken, { birthdate, name, zipCode, voteIdCard });
+
+    const user = User.fromJson(response.user);
+
+    yield put(updatedUserProfile({ user }));
+    yield put(savingProfile(false));
+    if (shouldNavigate) {
+      yield put(profileStateMachine());
+    }
+  } catch (e) {
+    logError(e, { tag: "updateProfile" });
+
+    yield put(savingProfile(false));
+
+    if (isUnauthorized(e)) return yield put(unauthorized({ type: "reset"}));
+
+    yield put(saveUserProfileError(e));
+  }
 }
 
 function* updateProfile({ mobileApi }) {
   yield takeLatest("PROFILE_UPDATE", function* ({ payload }) {
-    try {
-      const { birthdate, name, zipCode, voteIdCard } = payload;
+    const { birthdate, name, zipCode, voteIdCard, shouldNavigate } = payload;
 
-      yield put(savingProfile(true));
-
-      const authToken = yield select(currentAuthToken);
-      const response = yield call(mobileApi.updateProfile, authToken, { birthdate, name, zipCode, voteIdCard });
-
-      const user = User.fromJson(response.user);
-
-      yield put(updatedUserProfile({ user }));
-      yield put(savingProfile(false));
-      yield put(profileStateMachine());
-      yield call([Toast, Toast.show], locale.profileUpdated);
-    } catch (e) {
-      logError(e, { tag: "updateProfile" });
-
-      yield put(savingProfile(false));
-
-      if (isUnauthorized(e)) return yield put(unauthorized({ type: "reset"}));
-
-      yield put(saveUserProfileError(e));
-    }
+    yield call(updateProfileSaga, { mobileApi, birthdate, name, zipCode, voteIdCard, shouldNavigate })
   });
 }
 
@@ -420,6 +431,29 @@ function* validateProfile({ dispatch, mobileApi, walletStore }) {
   });
 }
 
+function* updateUser({ mobileApi }) {
+  yield takeLatest("UPDATE_USER", function* (action) {
+    const { avatar, birthdate, name, zipCode, currentPassword, newPassword } = action.payload.profile;
+    const { validAvatar, validProfileFields, validPassword } = action.payload.validations;
+    const shouldNavigate = false;
+
+    yield all([
+      validAvatar ? call(saveAvatarProfileSaga, { mobileApi, avatar, shouldNavigate }) : Promise.resolve(),
+      validProfileFields ? call(updateProfileSaga, { mobileApi, birthdate, name, zipCode, shouldNavigate }) : Promise.resolve(),
+      validPassword ? call(changePasswordSaga, { mobileApi, currentPassword, newPassword }) : Promise.resolve(),
+    ]);
+
+    const avatarError = yield select(avatarSaveErrors);
+    const profileErrors = yield select(profileSaveErrors);
+    const passwordErrors = yield select(getChangePasswordErrors);
+
+    if (isBlank(avatarError) && isBlank(profileErrors) && isBlank(passwordErrors)) {
+      yield put(navigate("plipsNav", { type: "reset" }));
+      yield call([Toast, Toast.show], locale.profileUpdated);
+    }
+  });
+}
+
 export default function* profileSaga({ dispatch, mobileApi, DeviceInfo, sessionStore, Crypto, walletStore }) {
   yield spawn(saveMainProfile, { mobileApi, sessionStore, Crypto });
   yield spawn(updateProfile, { mobileApi });
@@ -431,4 +465,5 @@ export default function* profileSaga({ dispatch, mobileApi, DeviceInfo, sessionS
   yield spawn(savePhoneProfile, { mobileApi, DeviceInfo });
   yield spawn(fetchProfileSaga, { mobileApi });
   yield fork(validateProfile, { dispatch, mobileApi, walletStore });
+  yield fork(updateUser, { mobileApi });
 }