+++ src/styles/avatar.js
@@ -12,5 +12,7 @@ export default StyleSheet.create({
   },
   container: {
     backgroundColor: "transparent",
+    alignItems: "center",
+    justifyContent: "center",
   },
 });